Move datalookup and transform spi to aho to separate folders
