+++ src/use-cases/dataLookup/index.ts
@@ -1,0 +1,3 @@
+export * from "./dataLookup"
+export { default as lookupPsaCodeByCrownCourtName } from "./lookupPsaCodeByCrownCourtName"
+export { default as lookupOrganisationUnitByCode } from "./lookupOrganisationUnitByCode"
