+++ src/use-cases/transformSpiToAho/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./transformSpiToAho"
