+++ src/use-cases/transformSpiToAho/populateCase.ts
