+++ src/use-cases/transformSpiToAho/populateDefendant.ts
@@ -1,6 +1,6 @@
 import type { Address, DefendantDetail, HearingDefendant } from "src/types/AnnotatedHearingOutcome"
 import type { ResultedCaseMessageParsedXml, SpiAddress, SpiCourtIndividualDefendant } from "src/types/IncomingMessage"
-import { lookupRemandStatusBySpiCode } from "./dataLookup"
+import { lookupRemandStatusBySpiCode } from "src/use-cases/dataLookup"
 import PopulateOffences from "./PopulateOffences"
 
 const formatPncIdentifier = (spiPNCIdentifier?: string): string | undefined =>
