+++ src/use-cases/transformSpiToAho/populateHearing.test.ts
@@ -1,5 +1,5 @@
 import { readFileSync } from "fs"
-import parseSpiResult from "./parseSpiResult"
+import parseSpiResult from "src/use-cases/parseSpiResult"
 import populateHearing from "./populateHearing"
 
 const message = readFileSync("test-data/input-message-001.xml", "utf-8")
