+++ src/use-cases/transformSpiToAho/populateHearing.ts
@@ -1,7 +1,7 @@
 import type { Hearing } from "src/types/AnnotatedHearingOutcome"
 import type { ResultedCaseMessageParsedXml } from "src/types/IncomingMessage"
 import removeSeconds from "src/utils/removeSeconds"
-import getOrganisationUnit from "./getOrganisationUnit"
+import getOrganisationUnit from "src/use-cases/getOrganisationUnit"
 
 export default (messageId: string, courtResult: ResultedCaseMessageParsedXml): Hearing => {
   const {
