+++ src/use-cases/transformSpiToAho/transformSpiToAho.ts
